requirements.txt
numpy
pandas
streamlit
groq
matplotlib
seaborn
python-dotenv
scikit-learn
# Streamlit app requirements

